% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomaRt.R
\name{randomaRt}
\alias{randomaRt}
\title{random_aRt}
\usage{
randomaRt(seed)
}
\arguments{
\item{seed}{A seed for reproducible function outputs}
}
\value{
An image made out of one of 5 aRtsy functions at random with a random colorpallette
}
\description{
Generates a random image
}
